+++ Mapping.py
@@ -1,45 +1,47 @@
-import numpy as np
-import pandas as pd
 import open3d as o3d
-import matplotlib.pyplot as plt
-import os
+import numpy as np
 import time
-
-folder_dir = "_out/"
-
-def read_ply(path):
-    cloud = o3d.io.read_point_cloud(path)
-    return cloud
 
 # create visualizer and window.
 vis = o3d.visualization.Visualizer()
 vis.create_window(height=480, width=640)
 
 # initialize pointcloud instance.
-pcd = (o3d.io.read_point_cloud("_out/00022637.ply"))
+pcd = o3d.geometry.PointCloud()
+# *optionally* add initial points
+points = np.random.rand(10, 3)
+pcd.points = o3d.utility.Vector3dVector(points)
 
 # include it in the visualizer before non-blocking visualization.
 vis.add_geometry(pcd)
 
 # to add new points each dt secs.
 dt = 0.01
+# number of points that will be added
+n_new = 10
+
 previous_t = time.time()
 
-for pl in os.listdir(folder_dir):
-    # check if the image ends with png
-    if (pl.endswith(".ply")):
+# run non-blocking visualization.
+# To exit, press 'q' or click the 'x' of the window.
+keep_running = True
+while keep_running:
 
-        if time.time() - previous_t > dt:
-            # Reading new points
-            path = folder_dir + pl
-            cloud = read_ply(path)
+    if time.time() - previous_t > dt:
+        # Options (uncomment each to try them out):
+        # 1) extend with ndarrays.
+        pcd.points.extend(np.random.rand(n_new, 3))
 
-            # extending with ndarrays.
-            pcd.points.extend(np.asarray(cloud.points))
+        # 2) extend with Vector3dVector instances.
+        # pcd.points.extend(
+        #     o3d.utility.Vector3dVector(np.random.rand(n_new, 3)))
 
-            # Updating the map
-            vis.update_geometry(pcd)
-            previous_t = time.time()
-        # Ploting the extended points
-        vis.update_renderer()
+        # 3) other iterables, e.g
+        # pcd.points.extend(np.random.rand(n_new, 3).tolist())
+
+        vis.update_geometry(pcd)
+        previous_t = time.time()
+
+    keep_running = vis.poll_events()
+    vis.update_renderer()
 vis.destroy_window()
